move string space kernel
